% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_catalogues.R
\name{sig_simulate_catalogues_from_signature}
\alias{sig_simulate_catalogues_from_signature}
\title{Simulate catalogues from a signature}
\usage{
sig_simulate_catalogues_from_signature(
  signature,
  n,
  n_catalogues,
  format = c("sigverse", "matrix"),
  prefix = "catalogue",
  seed = NULL,
  simplify = FALSE
)
}
\arguments{
\item{signature}{A signature object from the sigverse package. The signature
should contain \code{fraction} and \code{channel} columns. See \code{\link[sigshared:signature]{sigshared::example_signature()}}.}

\item{n}{The number of mutations to sample in each catalogue.}

\item{n_catalogues}{The number of catalogues to generate (default is 1).}

\item{format}{The format to return the generated catalogues in. Choose between:
\itemize{
\item "sigverse": Return a sigverse-style catalogue collection. See \code{\link[sigshared:catalogue_collections]{sigshared::example_catalogue_collection()}}.
\item "matrix": Return a matrix where each column corresponds to a simulated catalogue,
with rownames as channels and values as mutation counts.
}}

\item{prefix}{A string prefix to use for naming each catalogue. Defaults to "catalogue".}

\item{seed}{random seed to use for generation.}

\item{simplify}{A logical value indicating whether to return a single catalogue
(as a tibble) when only one catalogue is generated and \code{format = "sigverse"}.
Defaults to FALSE.}
}
\value{
Depending on the format:
\itemize{
\item If \code{format = "sigverse"}, a sigverse-style catalogue collection
(\code{\link[sigshared:catalogue_collections]{sigshared::example_catalogue_collection()}}) is returned.
\item If \code{format = "matrix"}, a matrix with one column per simulated catalogue
and rownames as mutation channels is returned.
\item If \code{simplify = TRUE} and only one catalogue is generated, a single catalogue
tibble (\code{\link[sigshared:catalogue]{sigshared::example_catalogue()}}) is returned.
}
}
\description{
This function performs multinomial sampling from a given mutational signature
to generate one or more mutation catalogues. The catalogues can be returned
in either 'sigverse' format or as a matrix.
}
\examples{
library(sigstash)
signatures <- sig_load('COSMIC_v3.4_SBS_GRCh38')
sig_simulate_catalogues_from_signature(signatures[["SBS1"]], n = 400, n_catalogues = 3)

}
