% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_catalogues.R
\name{sig_simulate_catalogues_from_signatures}
\alias{sig_simulate_catalogues_from_signatures}
\title{Simulate catalogues from multiple signatures}
\usage{
sig_simulate_catalogues_from_signatures(
  signatures,
  model,
  n,
  n_catalogues,
  format = c("sigverse", "matrix"),
  prefix = "catalogue",
  seed = NULL,
  simplify = FALSE
)
}
\arguments{
\item{signatures}{A list of signatures to combine and simulate from.
Each signature should be sigverse signature format (\code{\link[sigshared:signature_collection]{sigshared::example_signature_collection()}}).}

\item{model}{A model for combining the signatures. This can represent the
proportions or contributions of each signature in the combined result.}

\item{n}{The number of mutations to sample in each catalogue.}

\item{n_catalogues}{The number of catalogues to generate (default is 1).}

\item{format}{The format to return the generated catalogues in. Choose between:
\itemize{
\item "sigverse": Return a sigverse-style catalogue collection. See \code{\link[sigshared:catalogue_collections]{sigshared::example_catalogue_collection()}}.
\item "matrix": Return a matrix where each column corresponds to a simulated catalogue,
with rownames as channels and values as mutation counts.
}}

\item{prefix}{A string prefix to use for naming each catalogue. Defaults to "catalogue".}

\item{seed}{random seed to use for generation.}

\item{simplify}{A logical value indicating whether to return a single catalogue
(as a tibble) when only one catalogue is generated and \code{format = "sigverse"}.
Defaults to FALSE.}
}
\description{
This function perfectly combines multiple signatures in proportions described by the \code{model} argument,
then simulates catalogues based on multinomial sampling from the combined signature.
}
\details{
Note order of channels in resulting catalogue may differ from their order in
the signature collection due to sorting done by \code{\link[sigstats:sig_combine]{sigstats::sig_combine()}}.
}
\examples{
library(sigstash)
signatures <- sig_load('COSMIC_v3.4_SBS_GRCh38')
model <- c(SBS1 = 0.6, SBS5 = 0.4)
sig_simulate_catalogues_from_signatures(signatures, model, n = 400, n_catalogues=3)

}
