% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_catalogues.R
\name{sig_simulate_stratified}
\alias{sig_simulate_stratified}
\title{Simulate Signatures by Stratified Sampling}
\usage{
sig_simulate_stratified(
  signatures,
  model,
  n,
  n_catalogues,
  format = c("sigverse", "matrix"),
  prefix = "catalogue",
  seed = NULL,
  simplify = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{signatures}{A list of signatures to combine and simulate from.
Each signature should be sigverse signature format (\code{\link[sigshared:signature_collection]{sigshared::example_signature_collection()}}).}

\item{model}{A model for combining the signatures. This can represent the
proportions or contributions of each signature in the combined result.}

\item{n}{The number of mutations to sample in each catalogue.}

\item{n_catalogues}{The number of catalogues to generate (default is 1).}

\item{format}{The format to return the generated catalogues in. Choose between:
\itemize{
\item "sigverse": Return a sigverse-style catalogue collection. See \code{\link[sigshared:catalogue_collections]{sigshared::example_catalogue_collection()}}.
\item "matrix": Return a matrix where each column corresponds to a simulated catalogue,
with rownames as channels and values as mutation counts.
}}

\item{prefix}{A string prefix to use for naming each catalogue. Defaults to "catalogue".}

\item{seed}{random seed to use for generation.}

\item{simplify}{A logical value indicating whether to return a single catalogue
(as a tibble) when only one catalogue is generated and \code{format = "sigverse"}.
Defaults to FALSE.}

\item{verbose}{verbose mode (flag)}
}
\value{
A named list with:
\itemize{
\item \code{mutations_per_signature}: A named vector of how many mutations were sampled from each signature
\item \code{catalogues}: A list or matrix of simulated catalogues (see \code{format} argument)
}
}
\description{
Simulate catalogues by independently sampling from each signature in the model,
ensuring exact contribution proportions.
}
\examples{
library(sigstash)

#' # Load Signature Collection
signatures <- sig_load('COSMIC_v3.4_SBS_GRCh38')

# Define Model
model <- c(SBS1 = 0.6, SBS5 = 0.4)

# Independently sample 0.6x400 mutations from SBS1 & 0.4*400 mutations from SBS5
simulations <- sig_simulate_stratified(signatures, model, n = 400, n_catalogues=3)

# Confirm how many mutations were are from each signature
simulations$mutations_per_signature

simulations$catalogues

}
