% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigsim-package.R
\docType{package}
\name{sigsim-package}
\alias{sigsim}
\alias{sigsim-package}
\title{sigsim: Simulate Mutational Signature Catalogues for Benchmarking and Testing}
\description{
Tools for simulating mutation catalogues based on one or more mutational signatures, with or without added noise. Supports both stochastic sampling (via multinomial models) and exact stratified sampling to mimic real-world or controlled experimental conditions. Designed to aid benchmarking of signature detection, sensitivity analysis, and overfitting assessment in mutational signature analysis workflows.
}
\author{
\strong{Maintainer}: Sam El-Kamand \email{selkamand@ccia.org.au} (\href{https://orcid.org/0000-0003-2270-8088}{ORCID})

}
\keyword{internal}
